Fix more signed/unsigned mismatches uncovered by MSVC /W3
Reviewed by John.

No regression tests possible.

* platform/network/cf/ResourceResponseCFNet.cpp:
(WebCore::filenameHasSaneExtension):
* platform/win/FontDataWin.cpp:
(WebCore::FontData::containsCharacters):
* platform/win/PlatformScrollBarSafari.cpp:
(WebCore::):
* platform/win/UniscribeController.cpp:
(WebCore::UniscribeController::advance):
(WebCore::UniscribeController::shapeAndPlaceItem):
* platform/win/WebCoreTextRenderer.cpp:
(WebCore::doDrawTextAtPoint):
* plugins/win/PluginDatabaseWin.cpp:
(WebCore::parseVersionString):
* plugins/win/PluginPackageWin.cpp:
(WebCore::splitString):
* plugins/win/PluginViewWin.cpp:
(WebCore::freeStringArray):
(WebCore::PluginViewWin::invalidateTimerFired):


Canonical link: https://commits.webkit.org/18645@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@23877 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/platform/win/WebCoreTextRenderer.cpp
@@ -42,7 +42,7 @@ static void doDrawTextAtPoint(GraphicsContext& context, const String& text, cons
     font.drawText(&context, run, style, point);
 
     if (underlinedIndex >= 0) {
-        ASSERT(underlinedIndex < text.length());
+        ASSERT(underlinedIndex < static_cast<int>(text.length()));
 
         int beforeWidth;
         if (underlinedIndex > 0) {
